Add validations

+++ SmartSurveys.Core/Services/SurveyService.cs
@@ -2,6 +2,8 @@ using MapsterMapper;
 using SmartSurveys.Core.DAL.Repositories;
 using SmartSurveys.Core.DTO;
 using SmartSurveys.Core.Entities;
+using SmartSurveys.Core.Validators;
+using SmartSurveys.Core.Results;
 
 namespace SmartSurveys.Core.Services;
 
@@ -9,11 +11,15 @@ internal class SurveyService : ISurveyService
 {
     private readonly ISurveyRepository _surveyRepository;
     private readonly IMapper _mapper;
+    private readonly SurveyDtoValidator _surveyDtoValidator;
+    private readonly SurveyDetailsDtoValidator _surveyDetailsDtoValidator;
 
-    public SurveyService(ISurveyRepository surveyRepository, IMapper mapper)
+    public SurveyService(ISurveyRepository surveyRepository, IMapper mapper, SurveyDtoValidator surveyDtoValidator, SurveyDetailsDtoValidator surveyDetailsDtoValidator)
     {
         _surveyRepository = surveyRepository;
         _mapper = mapper;
+        _surveyDtoValidator = surveyDtoValidator;
+        _surveyDetailsDtoValidator = surveyDetailsDtoValidator;
     }
 
     public async Task<SurveyDetailsDto> GetAsync(int id)
@@ -30,22 +36,55 @@ internal class SurveyService : ISurveyService
         return _mapper.Map<IEnumerable<SurveyDto>>(surveys);
     }
 
-    public async Task CreateAsync(SurveyDetailsDto surveyDto)
+    public async Task<Result> CreateAsync(SurveyDetailsDto surveyDto)
     {
+        var validationResult = await _surveyDetailsDtoValidator.ValidateAsync(surveyDto);
+
+        if (!validationResult.IsValid)
+        {
+            return Result.Failure(validationResult.Errors.Select(e => e.ErrorMessage).ToList());
+        }
+        
         var survey = _mapper.Map<Survey>(surveyDto);
         
         await _surveyRepository.CreateAsync(survey);
+        
+        return Result.Success();
     }
 
-    public async Task UpdateAsync(SurveyDto surveyDto)
+    public async Task<Result> UpdateAsync(SurveyDto surveyDto)
     {
+        var exists = await _surveyRepository.ExistsAsync(surveyDto.Id);
+
+        if (!exists)
+        {
+            return Result.Failure("Survey does not exists.");
+        }
+        
+        var validationResult = await _surveyDtoValidator.ValidateAsync(surveyDto);
+
+        if (!validationResult.IsValid)
+        {
+            return Result.Failure(validationResult.Errors.Select(x => x.ErrorMessage).ToList());
+        }
+        
         var survey = _mapper.Map<Survey>(surveyDto);
 
         await _surveyRepository.UpdateAsync(survey);
+
+        return Result.Success();
     }
 
-    public async Task DeleteAsync(int id)
+    public async Task<Result> DeleteAsync(int id)
     {
+        var exists = await _surveyRepository.ExistsAsync(id);
+        
+        if (!exists)
+        {
+            return Result.Failure("Survey does not exists.");
+        }
+        
         await _surveyRepository.DeleteAsync(id);
+        return Result.Success();
     }
 }
